Implement function to combine single .fanmod files.
Single .fanmod files (storing graphs as a list of vertices) are combined into one big network that consists of all the single graphs. The file format of a .fanmod file is kept correctly.

It seems that the fanmod_linux command-line tool does not compute information about the random graphs correctly/at all. As a workaround one can open the combined networks file in the original fanmod software (only GUI) instead of the fanmod_linux command-line version (unoffical port of the original software).