Fix error where an empty fanmod file would crash the statistics calculation.
Trying to count the number of vertices in a PPI AA graph (.fanmod file) that is empty would lead to a value error and a NoneType as return value.
Now it simply returns zero, since counting vertices in an empty file is equal to counting no vertices at all.
Also added logging for that case.