Implement statistics for chemical properties of AA
Implement the calculation of statistics for chemical properties of amino
acids according to the 3 class and 5 class system, respectively.

Therefore, two '.pickle' files have been added, that store 20 amino
acids each and the corresponding chemical property according to the two
systems.
The software will parse 'aagraph_simple.id' files, that contain
information about amino acids contributing to PPIs. The chemical
properties of those amino acids are retrieved and the overall occurrence
of the chemical properties is counted.
The results will be saved as a bar chart and are included in the final
results HTML file.